src/req-agent/requirements.txt
python-dotenv==1.0.1
autogen-agentchat==0.4.0.dev13
autogen-ext[openai,azure]==0.4.0.dev13
markdown==3.7.0
PyGithub==2.5.0